refactor(cast): inline get logs bisecting implementation
